python_codes/helpers/requirements.txt
elasticsearch
pyyaml
unittest2
confluent_kafka
mysql.connector